Convert detail/edit pages to SSG for CDN cache delivery
Split 'use client' pages into server component wrapper + client component.
Added generateStaticParams to enable static generation (●) instead of
dynamic server rendering (ƒ), eliminating serverless cold starts on
page navigation.
