+++ docs/TASKS.md
@@ -83,6 +83,7 @@
 - [x] ローディング最低表示時間を1000ms→300msに短縮（体感速度改善 / 2026-03-23）
 - [x] PrismaClientを本番でもグローバル再利用に変更（コールドスタート時のDB再接続削減 / 2026-03-23）
 - [x] Supabaseクライアントをモジュールスコープで初期化に変更（`auth-server.ts`, `api/auth/admin` / 2026-03-23）
+- [x] 詳細・編集ページをSSG（静的生成）に変更しCDNキャッシュ配信に移行（`/report/[id]`, `/report/[id]/edit` / 2026-03-23）
 
 ## 残タスク
 - [x] allowedDevOrigins 警告の対応
